requirements.txt
torch
torchvision
accelerate
transformers
opencv-python
numpy
matplotlib
Pillow
tqdm